feat: (VTK-15) add endpoint de atualizar colab por ID.
