+++ VortekPonto/src/main/java/vortek/sistponto/VortekPonto/Repositories/ColaboradorRepository.java
@@ -1,0 +1,9 @@
+package vortek.sistponto.VortekPonto.Repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import vortek.sistponto.VortekPonto.Models.Colaborador;
+
+@Repository
+public interface ColaboradorRepository extends JpaRepository<Colaborador, Long> {
+}
