--- VortekPonto/src/main/java/vortek/sistponto/VortekPonto/Repositories/FuncionarioRepository.java
@@ -1,9 +1,0 @@
-package vortek.sistponto.VortekPonto.Repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import vortek.sistponto.VortekPonto.Models.Funcionario;
-
-@Repository
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
-}
